requirements.txt
alabaster @ file:///home/ktietz/src/ci/alabaster_1611921544520/work
alembic==1.6.5
appdirs==1.4.4
argh==0.26.2
arrow==1.1.0
astroid==2.5.7
async-generator @ file:///home/ktietz/src/ci/async_generator_1611927993394/work
atomicwrites==1.4.0
attrs @ file:///tmp/build/80754af9/attrs_1620827162558/work
autopep8 @ file:///tmp/build/80754af9/autopep8_1615918855173/work
Babel @ file:///tmp/build/80754af9/babel_1620871417480/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
bcrypt @ file:///C:/ci/bcrypt_1607022693089/work
binaryornot @ file:///tmp/build/80754af9/binaryornot_1617751525010/work
black==21.5b1
bleach @ file:///tmp/build/80754af9/bleach_1612211392645/work
blinker==1.4
brotlipy==0.7.0
bz2file==0.98
certifi==2021.5.30
cffi @ file:///C:/ci/cffi_1613247431367/work
chardet @ file:///C:/ci/chardet_1607706937985/work
click @ file:///tmp/build/80754af9/click_1621604852318/work
cloudpickle @ file:///tmp/build/80754af9/cloudpickle_1598884132938/work
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
cookiecutter==1.7.3
cryptography @ file:///C:/ci/cryptography_1616769504165/work
cycler==0.10.0
decorator @ file:///tmp/build/80754af9/decorator_1621259047763/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
diff-match-patch @ file:///tmp/build/80754af9/diff-match-patch_1594828741838/work
distro==1.5.0
dnspython==2.1.0
docutils @ file:///C:/ci/docutils_1620828264669/work
dominate==2.6.0
elasticsearch==7.13.1
email-validator==1.1.2
entrypoints==0.3
flake8 @ file:///tmp/build/80754af9/flake8_1615834841867/work
Flask==2.0.1
Flask-Admin==1.5.8
Flask-Babel==2.0.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.0.1
Flask-Moment==0.11.0
Flask-SQLAlchemy==2.5.1
Flask-Uploads==0.2.1
Flask-WTF==0.15.1
future @ file:///C:/ci/future_1607568713721/work
greenlet==1.1.0
idna @ file:///home/linux1/recipes/ci/idna_1610986105248/work
imagesize @ file:///home/ktietz/src/ci/imagesize_1611921604382/work
importlib-metadata==4.3.1
inflection==0.5.1
intervaltree @ file:///tmp/build/80754af9/intervaltree_1598376443606/work
ipykernel==5.5.5
ipython==7.24.0
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
isort @ file:///tmp/build/80754af9/isort_1616355431277/work
itsdangerous==2.0.1
jedi @ file:///C:/ci/jedi_1606932640618/work
Jinja2==3.0.1
jinja2-time @ file:///tmp/build/80754af9/jinja2-time_1617751524098/work
jsonschema @ file:///tmp/build/80754af9/jsonschema_1602607155483/work
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1616770841739/work
jupyter-core @ file:///C:/ci/jupyter_core_1612213515523/work
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
keyring @ file:///C:/ci/keyring_1621525114547/work
kiwisolver==1.3.1
lazy-object-proxy @ file:///C:/ci/lazy-object-proxy_1616529288960/work
Mako==1.1.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mccabe==0.6.1
minecart==0.3.0
mistune @ file:///C:/ci/mistune_1607359457024/work
mypy-extensions==0.4.3
nbclient @ file:///tmp/build/80754af9/nbclient_1614364831625/work
nbconvert @ file:///C:/ci/nbconvert_1607370555148/work
nbformat @ file:///tmp/build/80754af9/nbformat_1617383369282/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1613680548246/work
numpy==1.20.3
numpydoc @ file:///tmp/build/80754af9/numpydoc_1605117425582/work
olefile==0.46
opencv-python==4.5.2.54
packaging @ file:///tmp/build/80754af9/packaging_1611952188834/work
pandas==1.2.4
pandocfilters @ file:///C:/ci/pandocfilters_1605114832805/work
paramiko @ file:///tmp/build/80754af9/paramiko_1598886428689/work
parso==0.7.0
pathspec==0.8.1
pdfminer3k==1.3.4
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow @ file:///C:/ci/pillow_1617386320391/work
pluggy @ file:///C:/ci/pluggy_1615976440052/work
ply==3.11
poyo @ file:///tmp/build/80754af9/poyo_1617751526755/work
prompt-toolkit==3.0.18
psutil @ file:///C:/ci/psutil_1612298199233/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pycodestyle @ file:///home/ktietz/src/ci_mi/pycodestyle_1612807597675/work
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pydocstyle @ file:///tmp/build/80754af9/pydocstyle_1621600989141/work
pyflakes @ file:///home/ktietz/src/ci_ipy2/pyflakes_1612551159640/work
Pygments @ file:///tmp/build/80754af9/pygments_1621606182707/work
PyJWT==2.1.0
pylint==2.8.2
pyls-black @ file:///tmp/build/80754af9/pyls-black_1607553132291/work
pyls-spyder @ file:///tmp/build/80754af9/pyls-spyder_1613849700860/work
PyNaCl @ file:///C:/ci/pynacl_1607612759007/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
PyQt5==5.12.3
PyQt5-sip==12.9.0
PyQtWebEngine==5.12.1
pyrsistent @ file:///C:/ci/pyrsistent_1607359751294/work
PySocks @ file:///C:/ci/pysocks_1605307512533/work
python-dateutil @ file:///home/ktietz/src/ci/python-dateutil_1611928101742/work
python-dotenv==0.17.1
python-editor==1.0.4
python-jsonrpc-server @ file:///tmp/build/80754af9/python-jsonrpc-server_1600278539111/work
python-language-server @ file:///tmp/build/80754af9/python-language-server_1607972495879/work
python-slugify @ file:///tmp/build/80754af9/python-slugify_1620405669636/work
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
pywin32==301
pywin32-ctypes @ file:///C:/ci/pywin32-ctypes_1607553594546/work
PyYAML==5.4.1
pyzmq==22.1.0
QDarkStyle @ file:///tmp/build/80754af9/qdarkstyle_1617386714626/work
qstylizer==0.2.0
QtAwesome @ file:///tmp/build/80754af9/qtawesome_1615991616277/work
qtconsole==5.1.0
QtPy==1.9.0
regex==2021.4.4
requests @ file:///tmp/build/80754af9/requests_1608241421344/work
rope==0.19.0
Rtree @ file:///C:/ci/rtree_1618421015405/work
sip==4.19.13
six==1.16.0
snowballstemmer @ file:///tmp/build/80754af9/snowballstemmer_1611258885636/work
sortedcontainers==2.4.0
Sphinx==4.0.2
sphinxcontrib-applehelp @ file:///home/ktietz/src/ci/sphinxcontrib-applehelp_1611920841464/work
sphinxcontrib-devhelp @ file:///home/ktietz/src/ci/sphinxcontrib-devhelp_1611920923094/work
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath @ file:///home/ktietz/src/ci/sphinxcontrib-jsmath_1611920942228/work
sphinxcontrib-qthelp @ file:///home/ktietz/src/ci/sphinxcontrib-qthelp_1611921055322/work
sphinxcontrib-serializinghtml==1.1.5
spyder==5.0.3
spyder-kernels==2.0.3
SQLAlchemy==1.4.17
tabula-py==2.2.0
testpath==0.5.0
text-unidecode==1.3
textdistance @ file:///tmp/build/80754af9/textdistance_1612461398012/work
three-merge @ file:///tmp/build/80754af9/three-merge_1607553261110/work
tinycss @ file:///tmp/build/80754af9/tinycss_1617713798712/work
tinycss2==1.1.0
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
tornado @ file:///C:/ci/tornado_1606924294691/work
traitlets @ file:///home/ktietz/src/ci/traitlets_1611929699868/work
ujson @ file:///C:/ci/ujson_1611259568517/work
Unidecode @ file:///tmp/build/80754af9/unidecode_1614712377438/work
urllib3==1.26.5
visitor==0.1.3
watchdog @ file:///C:/ci/watchdog_1612471426646/work
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
Werkzeug==2.0.1
whichcraft @ file:///tmp/build/80754af9/whichcraft_1617751293875/work
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306162074/work
wincertstore==0.2
wrapt @ file:///C:/ci/wrapt_1607574570428/work
WTForms==2.3.3
yapf @ file:///tmp/build/80754af9/yapf_1615749224965/work
zipp @ file:///tmp/build/80754af9/zipp_1615904174917/work
